Add /app route and Debug Command

+++ src/app.res.ts
@@ -9,6 +9,7 @@ import { DB, group, groupMessage, user } from "./dbMagnament.res"
 import {ServerRunner} from "./server.res"
 import { v4 as uuidv4 } from 'uuid';
 
+const debug = process.env.NODE_ENV || false
 
 let load:Ora = ora(chalk.yellow("Esperando autenticación...")).start()
 let isLogged = false;
@@ -16,43 +17,23 @@ const settings = {
     sessionStorage: "./session",
     prefix: "?"
 }
+
+console.log(debug);
+
 let hasOwnSession;
 if (fs.existsSync(resolve(__dirname,settings.sessionStorage + ".json"))) {
     const token = require(`${resolve(__dirname,settings.sessionStorage + ".json")}`)
     hasOwnSession = token
 }
-
-const client = new Client({
-    session: hasOwnSession
-});
-
-client.on('qr', (qrCode) => {
-    // Generate and scan this code with your phone
-    isLogged = true
-    const code = qr.generate(qrCode, { small: true })
-    console.log(code)
-});
-
-
-client.on('message', async msg => {
-    const isGroup = (await msg.getChat()).isGroup;
-    const authorNumber = await msg.getContact();
-    load.text = "Nuevo Mensaje"
-    if(!isGroup){
-        savePrivate(msg,authorNumber);
-    }else{
-        saveGroup(msg,authorNumber)
-    }
-});
-
+    
 const saveGroup = async function(msg:Message, authorNumber:Contact){
+    let img = await (await msg.getContact()).getProfilePicUrl()
     let Group = (await msg.getChat())
     let groupName = Group.name;
 
     let number = authorNumber.id.user;
     let message = msg.body;
     let date = Date().toString();
-
     const pool = DB.GetConnection()
     let groupsPool = pool.get("groups");
     let group = groupsPool.find(e => e.groupName === groupName);
@@ -69,6 +50,7 @@ const saveGroup = async function(msg:Message, authorNumber:Contact){
     }else{
         const save:group = {
             groupName,
+            authorImage: img,
             messages: [
                 {
                     date: Date().toString(),
@@ -83,6 +65,7 @@ const saveGroup = async function(msg:Message, authorNumber:Contact){
 }
 
 const savePrivate = async  function(msg:Message,authorNumber:Contact){
+    let img = await (await msg.getContact()).getProfilePicUrl()
     const pool = DB.GetConnection()
     const userPool = pool.get("users");
     // @ts-ignore
@@ -133,7 +116,7 @@ const savePrivate = async  function(msg:Message,authorNumber:Contact){
         //@ts-ignore
         fs.writeFile(url, media.data, 'base64',error => console.log(error));
         save = {
-            
+            authorImage: img,
             authorNumber: authorNumber.id.user,
                 messages: [
             {
@@ -144,6 +127,7 @@ const savePrivate = async  function(msg:Message,authorNumber:Contact){
     }else{
         save = {
             authorNumber: authorNumber.id.user,
+            authorImage: img
                 messages: [
             {
                 date: Date().toString(),
@@ -157,30 +141,69 @@ const savePrivate = async  function(msg:Message,authorNumber:Contact){
     }
     load.succeed("Mensaje Guardado")
 }
-client.on("authenticated", e => {
-    hasOwnSession = e;
-    isLogged = true;
-    fs.writeFileSync(resolve(__dirname,settings.sessionStorage + ".json"), JSON.stringify(e));
-    load.succeed("Cliente conectado a Whatsapp!")
-    DB.CreateConnection(load)
-    ServerRunner.ora = load;
-    ServerRunner.listen()
-})
 
 
-client.on("auth_failture", e => {
-    load.stop();
-    console.error(e)
-    process.exit(1)
-})
 
-
-client.initialize();
-setTimeout(() => {
-    if (!isLogged) {
-        console.log(isLogged);
-        load.fail("Error tiempo de espera terminado");
-        fs.rmSync(resolve(__dirname,settings.sessionStorage + ".json"));
-        process.exit(1);
+const clientSettings = () => {
+    if(debug){
+        // Server and DB start
+        DB.CreateConnection(load)
+        ServerRunner.ora = load;
+        ServerRunner.listen()
+        return
     }
-}, 10000)
+    const client = new Client({
+        session: hasOwnSession
+    });
+
+    client.on('qr', (qrCode) => {
+        // Generate and scan this code with your phone
+        isLogged = true
+        const code = qr.generate(qrCode, { small: true })
+        console.log(code)
+    });
+
+
+    client.on('message', async msg => {
+
+        const isGroup = (await msg.getChat()).isGroup;
+        const authorNumber = await msg.getContact();
+        load.text = "Nuevo Mensaje"
+        if(!isGroup){
+            savePrivate(msg,authorNumber);
+        }else{
+            saveGroup(msg,authorNumber)
+        }
+    });
+
+    client.on("authenticated", e => {
+        hasOwnSession = e;
+        isLogged = true;
+        fs.writeFileSync(resolve(__dirname,settings.sessionStorage + ".json"), JSON.stringify(e));
+        load.succeed("Cliente conectado a Whatsapp!")
+        
+    })
+    
+    
+    client.on("auth_failture", e => {
+        load.stop();
+        console.error(e)
+        process.exit(1)
+    })
+
+    client.initialize();
+    setTimeout(() => {
+        if(debug){
+            return
+        }
+        if (!isLogged) {
+            console.log(isLogged);
+            load.fail("Error tiempo de espera terminado");
+            fs.rmSync(resolve(__dirname,settings.sessionStorage + ".json"));
+            process.exit(1);
+        }
+    }, 10000)
+}
+
+clientSettings()
+
